requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
cmake==3.28.1
    # via triton
filelock==3.13.1
    # via
    #   torch
    #   triton
idna==3.6
    # via requests
imageio==2.33.1
    # via scikit-image
jinja2==3.1.3
    # via torch
lazy-loader==0.3
    # via scikit-image
lit==17.0.6
    # via triton
markupsafe==2.1.3
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via
    #   scikit-image
    #   torch
numpy==1.26.3
    # via
    #   imageio
    #   scikit-image
    #   scipy
    #   tifffile
    #   torchvision
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   nvidia-cusolver-cu11
    #   torch
nvidia-cuda-cupti-cu11==11.7.101
    # via torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
nvidia-cufft-cu11==10.9.0.58
    # via torch
nvidia-curand-cu11==10.2.10.91
    # via torch
nvidia-cusolver-cu11==11.4.0.1
    # via torch
nvidia-cusparse-cu11==11.7.4.91
    # via torch
nvidia-nccl-cu11==2.14.3
    # via torch
nvidia-nvtx-cu11==11.7.91
    # via torch
packaging==23.2
    # via scikit-image
pillow==10.2.0
    # via
    #   imageio
    #   scikit-image
    #   torchvision
python-multipart==0.0.6
pyyaml==6.0.1
requests==2.31.0
    # via torchvision
scikit-image==0.22.0
scipy==1.11.4
    # via scikit-image
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
setuptools==69.2.0
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-cupti-cu11
    #   nvidia-cuda-runtime-cu11
    #   nvidia-curand-cu11
    #   nvidia-cusparse-cu11
    #   nvidia-nvtx-cu11
sympy==1.12
    # via torch
tifffile==2023.12.9
    # via scikit-image
torch==2.0.0
    # via
    #   torchvision
    #   triton
torchvision==0.15.1
triton==2.0.0
    # via torch
typing-extensions==4.9.0
    # via torch
urllib3==2.1.0
    # via requests
wheel==0.42.0
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-cupti-cu11
    #   nvidia-cuda-runtime-cu11
    #   nvidia-curand-cu11
    #   nvidia-cusparse-cu11
    #   nvidia-nvtx-cu11
jupyter
matplotlib
tqdm